requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
ddgs>=9.0.0
